sbi_spec
Module base

Chapter 4. Base Extension (EID #0x10).

Modules

impl_id

SBI Implementation IDs.

Structs

Version

SBI specification version.

Constants

EID_BASE

Extension ID for RISC-V SBI Base extension.

GET_MARCHID

Function ID to get the value of marchid register in the current environment.

GET_MIMPID

Function ID to get the value of mimpid register in the current environment.

GET_MVENDORID

Function ID to get the value of mvendorid register in the current environment.

GET_SBI_IMPL_ID

Function ID to get the current SBI implementation ID.

GET_SBI_IMPL_VERSION

Function ID to get the current SBI implementation version.

GET_SBI_SPEC_VERSION

Function ID to get the current SBI specification version.

PROBE_EXTENSION

Function ID to probe information about one SBI extension from the current environment.

UNAVAILABLE_EXTENSION

Default probe value for the target SBI extension is unavailable.