Struct sbi_spec::binary::SbiRet

#[repr(C)]
pub struct SbiRet {
    pub error: usize,
    pub value: usize,
}

SBI functions return type

SBI functions must return a pair of values in a0 and a1, with a0 returning an error code. This is analogous to returning the C structure SbiRet.

Note: if this structure is used in function return on conventional Rust code, it would not require to pin memory representation as extern C. The repr(C) is set in case that some users want to use this structure in FFI code.

Fields
error: usize

Error number

value: usize

Result value

Implementations

impl SbiRet

pub const fn success(value: usize) -> Self

Returns success SBI state with given value.

pub const fn failed() -> Self

The SBI call request failed for unknown reasons.

pub const fn not_supported() -> Self

SBI call failed due to not supported by target ISA, operation type not supported, or target operation type not implemented on purpose.

pub const fn invalid_param() -> Self

SBI call failed due to invalid hart mask parameter, invalid target hart id, invalid operation type, or invalid resource index.

pub const fn invalid_address() -> Self

SBI call failed for invalid mask start address, not a valid physical address parameter, or the target address is prohibited by PMP to run in supervisor mode.

pub const fn already_available() -> Self

SBI call failed for the target resource is already available, e.g. the target hart is already started when caller still request it to start.

pub const fn already_started() -> Self

SBI call failed for the target resource is already started, e.g. target performance counter is started.

pub const fn already_stopped() -> Self

SBI call failed for the target resource is already stopped, e.g. target performance counter is stopped.

impl SbiRet

pub const fn into_result(self) -> Result<usize, Error>

Converts to a Result of value and error.

pub const fn is_ok(&self) -> bool

Returns true if current SBI return succeeded.

pub const fn is_err(&self) -> bool

Returns true if current SBI return is an error.

pub fn ok(self) -> Option<usize>

Converts from SbiRet to Option<usize>.

pub fn err(self) -> Option<Error>

Converts from SbiRet to Option<Error>.

pub fn map<U, F: FnOnce(usize) -> U>(self, op: F) -> Result<U, Error>

Maps a SbiRet to Result<U, E> by applying a function to a contained success value, leaving an error value untouched.

pub fn map_or<U, F: FnOnce(usize) -> U>(self, default: U, f: F) -> U

Returns the provided default (if error), or applies a function to the contained value (if success),

pub fn map_or_else<U, D: FnOnce(Error) -> U, F: FnOnce(usize) -> U>( self, default: D, f: F ) -> U

Maps a SbiRet to usize value by applying fallback function default to a contained error, or function f to a contained success value.

pub fn map_err<F, O: FnOnce(Error) -> F>(self, op: O) -> Result<usize, F>

Maps a SbiRet to Result<T, F> by applying a function to a contained error as Error struct, leaving success value untouched.

pub fn expect(self, msg: &str) -> usize

Returns the contained success value, consuming the self value.

Panics

Panics if self is an SBI error with a panic message including the passed message, and the content of the SBI state.

pub fn unwrap(self) -> usize

Returns the contained success value, consuming the self value.

Panics

Panics if self is an SBI error, with a panic message provided by the SBI error converted into Error struct.

pub fn expect_err(self, msg: &str) -> Error

Returns the contained error as Error struct, consuming the self value.

Panics

Panics if the self is SBI success value, with a panic message including the passed message, and the content of the success value.

pub fn unwrap_err(self) -> Error

Returns the contained error as Error struct, consuming the self value.

Panics

Panics if the self is SBI success value, with a custom panic message provided by the success value.

pub fn and(self, res: Result<usize, Error>) -> Result<usize, Error>

Returns res if self is success value, otherwise otherwise returns the contained error of self as Error struct.

pub fn and_then<U, F: FnOnce(usize) -> Result<U, Error>>( self, op: F ) -> Result<U, Error>

Calls op if self is success value, otherwise returns the contained error as Error struct.

pub fn or<F>(self, res: Result<usize, F>) -> Result<usize, F>

Returns res if self is SBI error, otherwise returns the success value of self.

pub fn or_else<F, O: FnOnce(Error) -> Result<usize, F>>( self, op: O ) -> Result<usize, F>

Calls op if self is SBI error, otherwise returns the success value of self.

pub fn unwrap_or(self, default: usize) -> usize

Returns the contained success value or a provided default.

pub fn unwrap_or_else<F: FnOnce(Error) -> usize>(self, op: F) -> usize

Returns the contained success value or computes it from a closure.