sbi_spec
Module dbcn

Chapter 12. Debug Console Extension (EID #0x4442434E “DBCN”).

Constants

CONSOLE_READ

Function ID to read bytes from the debug console into an output memory.

CONSOLE_WRITE

Function ID to write bytes to the debug console from input memory.

CONSOLE_WRITE_BYTE

Function ID to write a single byte to the debug console.

EID_DBCN

Extension ID for Debug Console Extension.