sbi_spec
Module hsm

Chapter 9. Hart State Management Extension (EID #0x48534D “HSM”).

Modules

hart_state

Hart states.

suspend_type

Hart suspend types.

Constants

EID_HSM

Extension ID for Hart State Management extension.

HART_GET_STATUS

Function ID to get the current status (or HSM state id) of the given hart.

HART_START

Function ID to start executing the given hart at specified address in supervisor-mode.

HART_STOP

Function ID to stop executing the calling hart in supervisor-mode.

HART_SUSPEND

Function ID to put the calling hart into suspend or platform-specific lower power states.