sbi_spec
Module nacl

Chapter 15. Nested Acceleration Extension (EID #0x4E41434C “NACL”).

Modules

feature_id

Nested Acceleration Feature ID.

shmem_size

Size of shared memory set by supervisor software for current hart.

Constants

EID_NACL

Extension ID for Nested Acceleration Extension.

PROBE_FEATURE

Function ID to probe a nested acceleration feature.

SET_SHMEM

Function ID to set and enable the shared memory for nested acceleration on the calling hart.

SYNC_CSR

Function ID to synchronize CSRs in the nested acceleration shared memory.

SYNC_HFENCE

Function ID to synchronize HFENCEs in the nested acceleration shared memory.

SYNC_SRET

Function ID to synchronize CSRs and HFENCEs in the nested acceleration shared memory and emulate the SRET instruction.