sbi_spec::nacl
Module shmem_size

Size of shared memory set by supervisor software for current hart.

NACL shared memory includes scratch space and CSR space. Due to the difference of CSR width, this size varies between different XLEN values. NATIVE constant here only matches the integer width for the target this crate is compiled. If you are writing an SEE with different XLEN from the host platform, you should choose other correct constant value from RV32, RV64 or RV128 in module shmem_size instead.

Constants

NATIVE

Size of NACL shared memory on platforms with XLEN of the same width as the current platform.

RV32

Size of NACL shared memory on RV32 platforms.

RV64

Size of NACL shared memory on RV64 platforms.

RV128

Size of NACL shared memory on RV128 platforms.