sbi_spec
Module pmu

Chapter 11. Performance Monitoring Unit Extension (EID #0x504D55 “PMU”).

Modules

cache_event

Hardware Cache Event ID.

cache_operation

Hardware Cache Operation ID.

cache_result

Hardware Cache Operation Result ID.

event_type

PMU Event Types.

firmware_event

Firmware Event Codes.

hardware_event

Hardware General Event Codes.

Constants

COUNTER_CONFIG_MATCHING

Function ID to find and configure a counter from a set of counters.

COUNTER_FW_READ

Function ID to provide the current value of a firmware counter.

COUNTER_FW_READ_HI

Function ID to provide the upper 32 bits from the value of the current firmware counter.

COUNTER_GET_INFO

Function ID to get details about the specified counter.

COUNTER_START

Function ID to start or enable a set of counters on the calling hart with the specified initial value.

COUNTER_STOP

Function ID to stop or disable a set of counters on the calling hart.

EID_PMU

Extension ID for Performance Monitoring Unit extension.

NUM_COUNTERS

Function ID to get the number of counters, both hardware and firmware.

SNAPSHOT_SET_SHMEM

Function ID to set and enable the PMU snapshot shared memory.